ContextFreeSQL/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile pyproject.toml
#
sqlparse==0.5.2
    # via ContextFreeSQL (pyproject.toml)
sqlglot
psycopg2-binary
python-dotenv
pandas
numpy
pydantic